pyproject.toml
[project]
name = "langgraph-deployment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckduckgo-search>=8.0.2",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.60",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.4",
    "langgraph>=0.4.5",
    "langgraph-api>=0.2.30",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "sqlalchemy>=2.0.41",
    "together>=1.5.8",
]

[dependency-groups]
dev=[
    "ipykernel>=6.29.5",
    "langgraph-cli>=0.2.10",
]
